backend/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
ecdsa==0.18.0
fastapi==0.85.1
fastapi-jwt-auth==0.5.0
flake8==6.0.0
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
inflect==6.0.0
mccabe==0.7.0
paramiko==3.2.0
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pyflakes==3.0.1
PyJWT==1.7.1
PyNaCl==1.5.0
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
scp==0.14.5
six==1.16.0
sniffio==1.3.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.15
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
